_grader code complete. Time for testing! Fixed the submission form to not include non-user-supplied fields

+++ collector/views.py
@@ -2,7 +2,10 @@
 from django.shortcuts import render_to_response, get_object_or_404
 from django.http import HttpResponseRedirect, Http404
 
+from django.core.files import File
+
 from assignmentcollectorgrader.collector.models import *
+from assignmentcollectorgrader.settings import JUNIT_ROOT
 
 def render_to_csrf(request, template, context):
     from django.core.context_processors import csrf
@@ -61,10 +64,12 @@ def submit_assignment(request, year, term, course_id, assn_name):
                 late = "Your submission is past the due date.\n"
             else:
                 late = ""
+            # Calculate which submission number this is    
+            
             # either warn the user they have submitted a late assignment
             # or in the validation, check if it is late, and disallow submission
             # then report an error to the user
-            form.save()
+            submission = form.save()
             
             # Once saved, run the tests on the uploaded assignment and save the output as a string
             
@@ -83,7 +88,7 @@ def submit_assignment(request, year, term, course_id, assn_name):
     
 
 def _grader(assignment, submission):
-    import os.path, tempfile, zipfile, re, shutil
+    import os.path, os, tempfile, zipfile, re, shutil, subprocess, time
     # create a temporary directory
     temp = tempfile.mkdtemp()
     # extract the student's work into the temporary directory
@@ -99,15 +104,63 @@ def _grader(assignment, submission):
     ## extract legal files to the temporary directory
     jar.extractall(temp, to_extract) 
     # delete all class files
+    class_files = os.listdir(temp)
     ## find all class files
     ## delete them
+    for i in class_files:
+        if re.search("\.class$", i):
+            os.remove(i)
     # if the grader is a java file, copy it to the temporary dir
+    shutil.copy2(assignment.test_file.path, temp)
     # if the grader is a JAR, extract the grader into the temporary dir
     # compile all java files, saving all output
+    output_handle, output_path = tempfile.mkstemp(suffix=".log", dir=temp, text=True)
     ## add the location of JUnit to the classpath, or environment
+    if 'CLASSPATH' in os.environ:
+        os.environ['CLASSPATH'] = os.environ['CLASSPATH'] + os.pathsep + JUNIT_ROOT
+    else:
+        os.environ['CLASSPATH'] = JUNIT_ROOT
     ## actually compile
-    # run the tests, saving output
-    # delete the temporary directory
-    shutil.rmtree(temp)
-    # return the output
+    args = ['javac',  '*.java'] # '-verbose',
+    javac = subprocess.Popen(args=args, shell=False, stdout=output_handle, stderr=subprocess.STDOUT, cwd=temp, )
+    ## wait until the compilation is done
+    javac.wait() # may deadlock. Let's see
+    #while javac.poll() is None:
+    #    time.sleep(.25) # this is a work-around to avoid using javac.wait() because of the risk of deadlocking the subprocess
+    # If compilation is unsuccessful, return the program output
+    if javac.returncode > 0:
+    ## first move the output file to the submission directory tree
+        #submission.grade_log.save(submission.fileurl(os.path.basename(output_path)), File(open(output_path)), save=True)
+        submission.grade_log.save(os.path.basename(output_path), File(open(output_path)), save=True)  
+    ## read in the output
+        output = submission.grade_log.read()
+    ## close open files
+        os.close(output_handle)
+        jar.close()
+    ## delete the temporary folder
+        shutil.rmtree(temp)
+    ## return the output
+        return output
+    # else, run the tests, saving output
+    else: 
+    ## Run the tests
+        name, ext = os.path.splitext(os.path.basename(assignment.test_file))
+        args = ['java', 'junit.textui.TestRunner', name]
+        java = subprocess.Popen(args=args, shell=False, stdout=output_handle, stderr=subprocess.STDOUT, cwd=temp, )
+    ## Wait for the tests to complete
+        java.wait()
+        #while java.poll() is None:
+        #    pass
+    ## Save the test output
+        submission.grade_log.save(os.path.basename(output_path), File(open(output_path)), save=True)  
+    ## read in the output
+        output = submission.grade_log.read()
+    ## close open files
+        os.close(output_handle)
+        jar.close()
+    ## delete the temporary directory
+        shutil.rmtree(temp)
+    ## return the output
+        return output
+
